% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{discrete_palettes}
\alias{discrete_palettes}
\title{Color palettes and information}
\format{
An object of class \code{list} of length 35.
}
\usage{
discrete_palettes
}
\description{
Color palettes and information
}
\keyword{internal}
