% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\name{evenly_sample_colors}
\alias{evenly_sample_colors}
\title{Evenly sample colors from a palette}
\usage{
evenly_sample_colors(colors, n)
}
\arguments{
\item{colors}{Character vector of colors}

\item{n}{Integer, number of colors to return}
}
\value{
Character vector of selected colors
}
\description{
Evenly sample colors from a palette
}
\keyword{internal}
