% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\name{generate_additional_colors}
\alias{generate_additional_colors}
\title{Generate additional colors that are maximally distinct}
\usage{
generate_additional_colors(existing_colors, n_additional)
}
\arguments{
\item{existing_colors}{Character vector of existing hex colors}

\item{n_additional}{Integer, number of additional colors needed}
}
\value{
Character vector of new hex colors
}
\description{
Generate additional colors that are maximally distinct
}
\keyword{internal}
