% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{get_continuous_palette}
\alias{get_continuous_palette}
\title{Get continuous palette}
\usage{
get_continuous_palette(palette_name, n = 100, reverse = FALSE)
}
\arguments{
\item{palette_name}{Character, name of the palette}

\item{n}{Integer, number of colors to return}

\item{reverse}{Logical, whether to reverse the palette}
}
\value{
Character vector of colors
}
\description{
Get continuous palette
}
\examples{
get_continuous_palette("viridis", n = 100)
}
