% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{get_discrete_palette_cr}
\alias{get_discrete_palette_cr}
\title{Get discrete palette}
\usage{
get_discrete_palette_cr(palette_name, n = NULL, reverse = FALSE)
}
\arguments{
\item{palette_name}{Character, name of the palette (can be either "name" or "source.name" format)}

\item{n}{Integer, number of colors to return}

\item{reverse}{Logical, whether to reverse the palette}
}
\value{
Character vector of colors
}
\description{
Get discrete palette
}
\examples{
get_discrete_palette_cr("friendly", n = 4)
get_discrete_palette_cr("tidyplots.friendly", n = 4)
}
