% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_functions.R
\name{get_diverging_palette_cr}
\alias{get_diverging_palette_cr}
\title{Get diverging palette}
\usage{
get_diverging_palette_cr(palette_name, n = 100, reverse = FALSE)
}
\arguments{
\item{palette_name}{Character, name of the palette (can be either "name" or "source.name" format)}

\item{n}{Integer, number of colors to return}

\item{reverse}{Logical, whether to reverse the palette}
}
\value{
Character vector of colors
}
\description{
Get diverging palette
}
\examples{
get_diverging_palette_cr("spectral", n = 100)
get_diverging_palette_cr("tidyplots.spectral", n = 100)
}
