% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palette_info}
\alias{palette_info}
\title{Color palette information}
\format{
A data frame with columns:
\describe{
  \item{palette_name}{Full name of the palette (source.name)}
  \item{source}{Source/origin of the palette}
  \item{name}{Name of the palette}
  \item{type}{Type of the palette: discrete, continuous, or diverging}
  \item{n_colors}{Number of colors in the palette}
  \item{colors}{List column containing the color hex codes}
}
}
\usage{
palette_info
}
\description{
A dataset containing information about all available color palettes in the package
}
\keyword{datasets}
