% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_functions.R
\name{scale_color_continuous_cr}
\alias{scale_color_continuous_cr}
\title{Scale color for continuous palettes}
\usage{
scale_color_continuous_cr(
  palette_name = "tidyplots.inferno",
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette_name}{Character, name of the palette (can be either "name" or "source.name" format)}

\item{reverse}{Logical, whether to reverse the palette}

\item{...}{Additional arguments passed to scale_color_gradientn}
}
\value{
A ggplot2 continuous color scale
}
\description{
Scale color for continuous palettes
}
