% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_functions.R
\name{scale_color_discrete_cr}
\alias{scale_color_discrete_cr}
\title{Scale color for discrete palettes}
\usage{
scale_color_discrete_cr(
  palette_name = "tidyplots.friendly",
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette_name}{Character, name of the palette (can be either "name" or "source.name" format)}

\item{reverse}{Logical, whether to reverse the palette}

\item{...}{Additional arguments passed to discrete_scale}
}
\value{
A ggplot2 discrete color scale
}
\description{
Scale color for discrete palettes
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point() +
  scale_color_discrete_cr("friendly")

# Or use the full name
ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point() +
  scale_color_discrete_cr("tidyplots.friendly")
}
