% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_functions.R
\name{scale_fill_continuous_pal}
\alias{scale_fill_continuous_pal}
\title{Scale fill for continuous palettes}
\usage{
scale_fill_continuous_pal(palette_name = "inferno", reverse = FALSE, ...)
}
\arguments{
\item{palette_name}{Character, name of the palette}

\item{reverse}{Logical, whether to reverse the palette}

\item{...}{Additional arguments passed to scale_fill_gradientn}
}
\description{
Scale fill for continuous palettes
}
