% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_functions.R
\name{scale_fill_discrete_cr}
\alias{scale_fill_discrete_cr}
\title{Scale fill for discrete palettes}
\usage{
scale_fill_discrete_cr(
  palette_name = "tidyplots.friendly",
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette_name}{Character, name of the palette (can be either "name" or "source.name" format)}

\item{reverse}{Logical, whether to reverse the palette}

\item{...}{Additional arguments passed to discrete_scale}
}
\value{
A ggplot2 discrete fill scale
}
\description{
Scale fill for discrete palettes
}
