% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_functions.R
\name{scale_fill_discrete_pal}
\alias{scale_fill_discrete_pal}
\title{Scale fill for discrete palettes}
\usage{
scale_fill_discrete_pal(palette_name = "friendly", reverse = FALSE, ...)
}
\arguments{
\item{palette_name}{Character, name of the palette}

\item{reverse}{Logical, whether to reverse the palette}

\item{...}{Additional arguments passed to discrete_scale}
}
\description{
Scale fill for discrete palettes
}
