% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_functions.R
\name{scale_fill_diverging_pal}
\alias{scale_fill_diverging_pal}
\title{Scale fill for diverging palettes}
\usage{
scale_fill_diverging_pal(
  palette_name = "spectral",
  midpoint = 0,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette_name}{Character, name of the palette}

\item{midpoint}{Numeric, the midpoint for diverging scales}

\item{reverse}{Logical, whether to reverse the palette}

\item{...}{Additional arguments passed to scale_fill_gradientn}
}
\description{
Scale fill for diverging palettes
}
