% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_functions.R
\name{bold_fonts}
\alias{bold_fonts}
\title{Apply bold fonts to theme elements}
\usage{
bold_fonts(
  base_size = 11,
  base_face = "plain",
  title_face = "bold",
  subtitle_face = "plain",
  axis_title_face = "bold",
  axis_text_face = "plain",
  legend_title_face = "bold",
  legend_text_face = "plain"
)
}
\arguments{
\item{base_size}{Base font size}

\item{base_face}{Base font face}

\item{title_face}{Title font face}

\item{subtitle_face}{Subtitle font face}

\item{axis_title_face}{Axis title font face}

\item{axis_text_face}{Axis text font face}

\item{legend_title_face}{Legend title font face}

\item{legend_text_face}{Legend text font face}
}
\description{
Apply bold fonts to theme elements
}
