% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_functions.R
\name{bold_fonts}
\alias{bold_fonts}
\title{Title}
\usage{
bold_fonts(face = "bold", title_size_ratio = 1.15, legend_text_face = "plain")
}
\arguments{
\item{face}{fontface for all text}

\item{title_size_ratio}{relative size to original of all titles (plots, axis, legends, strips)}

\item{legend_face}{fontface for legend text}
}
\description{
Title
}
