% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_functions.R
\name{scale_fill_continuous_pal}
\alias{scale_fill_continuous_pal}
\title{Scale fill for continuous palettes}
\usage{
scale_fill_continuous_pal(
  palette_name = "tidyplots.inferno",
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette_name}{Character, name of the palette (can be either "name" or "source.name" format)}

\item{reverse}{Logical, whether to reverse the palette}

\item{...}{Additional arguments passed to scale_fill_gradientn}
}
\description{
Scale fill for continuous palettes
}
